requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
mlflow>=2.8.0
dvc>=3.0.0
pyyaml>=6.0
joblib>=1.2.0